Many improvements.

+++ nodelistj-download-nodelists/src/main/java/ru/gavrilovegor519/nodelistj_download_nodelists/util/UpdateNodelists.java
@@ -13,6 +13,8 @@ import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.time.Year;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.concurrent.TimeUnit;
 
 @RequiredArgsConstructor
@@ -21,7 +23,7 @@ import java.util.concurrent.TimeUnit;
 public class UpdateNodelists {
     private final MinioClient minioClient;
     private final FtpClient ftpClient;
-    private final KafkaTemplate<Void, Boolean> kafkaTemplate;
+    private final KafkaTemplate<String, List<String>> kafkaTemplate;
 
     @Value("${ftp.path}")
     private String ftpPath;
@@ -34,24 +36,24 @@ public class UpdateNodelists {
         try {
             createBucket();
             ftpClient.open();
-            boolean downloaded = false;
 
             for (int i = Year.now().getValue(); i >= downloadFromYear; i--) {
+                List<String> objects = new ArrayList<>();
                 String[] listFiles = ftpClient.listFiles(ftpPath + i);
 
                 for (String file : listFiles) {
-                    if (!isObjectExist(file)) {
+                    if (file.matches(ftpPath + "\\d{4}/nodelist\\.\\d{3}") && !isObjectExist(file)) {
                         try (ByteArrayOutputStream byteArrayOutputStream = ftpClient.downloadFile(file)) {
                             minioClient.putObject(PutObjectArgs.builder().bucket("nodehist").object(file)
                                     .stream(new ByteArrayInputStream(byteArrayOutputStream.toByteArray()),
                                             byteArrayOutputStream.size(), -1).build());
-                            downloaded = true;
+                            objects.add(file);
                         }
                     }
                 }
 
-                if (downloaded) {
-                    kafkaTemplate.send("download_nodelists_is_finished_topic", true);
+                if (!objects.isEmpty()) {
+                    kafkaTemplate.send("download_nodelists_is_finished_topic", objects);
                 }
             }
         } catch (Exception e) {
